rename package machines to machine
