+++ src/main/kotlin/com/github/trcdeveloppers/clayium/common/ClayiumCommonProxy.kt
@@ -2,7 +2,7 @@ package com.github.trcdeveloppers.clayium.common
 
 import com.github.trcdeveloppers.clayium.Clayium
 import com.github.trcdeveloppers.clayium.common.blocks.ClayiumBlocks
-import com.github.trcdeveloppers.clayium.common.blocks.machines.clayworktable.TileClayWorkTable
+import com.github.trcdeveloppers.clayium.common.blocks.machine.clayworktable.TileClayWorkTable
 import com.github.trcdeveloppers.clayium.common.items.ClayiumItems
 import com.github.trcdeveloppers.clayium.common.worldgen.ClayOreGenerator
 import net.minecraft.block.Block
